+++ src/zdb.h
@@ -14,9 +14,9 @@
     void zdbfs_zdb_free(zdbfs_t *fs);
 
     int zdb_flush(redisContext *remote);
-    zdb_reply_t *zdb_get(redisContext *remote, uint32_t id);
-    uint32_t zdb_set(redisContext *remote, uint32_t id, const void *buffer, size_t length);
-    int zdb_del(redisContext *remote, uint32_t id);
+    zdb_reply_t *zdb_get(redisContext *remote, uint64_t id);
+    uint64_t zdb_set(redisContext *remote, uint64_t id, const void *buffer, size_t length);
+    int zdb_del(redisContext *remote, uint64_t id);
     zdb_nsinfo_t *zdb_nsinfo(redisContext *remote, char *namespace);
 
     void zdbfs_zdb_reply_free(zdb_reply_t *reply);
